Move standalone StoryChunk extract/ingest test case to test/integration/Keeper-Grapher
Rename all collector to grapher
