+++ test/integration/CMakeLists.txt
@@ -1,3 +1,4 @@
 cmake_minimum_required(VERSION 3.19)
 
 add_subdirectory(Client)
+add_subdirectory(Keeper-Grapher)
